Moved some functions in one library. Deleted trash classess
