+++ RemAI/HelpCallers.cs
@@ -5,8 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using RemAI.utils;
-
+using Controller;
 namespace RemAI
 {
    internal class HelpCallers : RemAI
@@ -41,7 +40,7 @@ namespace RemAI
             Console.WriteLine("--- Information about program ---\n");
             Console.WriteLine(" Name: RemAI (bot) ");
             Console.WriteLine(" Type: Legacy ");
-            Console.WriteLine(" Permission: " + Utility.CheckPermission());
+            Console.WriteLine(" Permission: " + Control.CheckPermission() );
             Console.WriteLine(" Version: 1.2p");
             Console.WriteLine(" You can check updates on -> https://github.com/VladyslavHimself/RemAI/releases");
 
@@ -72,17 +71,6 @@ namespace RemAI
             Console.WriteLine("-exit - exit program\n");
 
         }
-
-        public static void ClearConsole()
-        {
-
-            Console.WriteLine("Clear the console...");
-            Thread.Sleep(100);
-            Console.Clear();
-
-        }
-
-    
 
         public static bool UpToDev(bool isDeveloper, string devkey)
         {
