--- RemAI/Utility.cs
@@ -1,20 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace RemAI.utils
-{
-    internal class Utility : RemAI
-    {
-        public static string CheckPermission()
-        {
-
-                    // #TODO Utility checker permission in session.
-            return "unknown";
-        }
-
-
-    }
-}
